requirements.txt
appdirs==1.4.3
Django==1.11.2
django-appconf==1.0.2
django-imagekit==4.0.1
django-markdownx==2.0.21
django-material==1.0.0
Markdown==2.6.8
olefile==0.44
packaging==16.8
pilkit==2.0
Pillow==4.1.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
